Implement time series logfile var records
Includes limited pretty-printing of the record to a JSON file.
Currently, arrays are broken up with one value per line, which is in
accordance with normal JSON practices yet rather annoying in this
instance.  Ideally, arrays would stay on one line, but at least the time
series arrays seem to end up at the end of the JSON file.  The section
for instantaneous variables at the beginning of the JSON file remains
quite readable, so it's good enough for now.