Get rid of step parasite axes on var vs. time plot
In an FEA analysis, time step ID & time are not guaranteed to linear
transformations of each other, due to varied time step intervals, so
they should not be plotted together.

If a var vs. step plot is desired, it should be prepared as a separate
plot from var vs. time.