Flatten analysis directory structure
Make the output files easier to browse by putting similar files for all
cases at the same level in the analysis directory tree.

`tabulate_case_write` was modified so that by default it writes to the
same directory that contains the case's FEBio files.  In a multi-case
analysis, the analysis directory structure is now controlled by
`tabulate_case_write`'s caller.  This is a clearer separation of
concerns.