Don't write multiple <plotfile> elements
The sensitivity analysis case generation re-uses the same XML tree for
each case.  Prior to this commit, it added a new <plotfile> element to
the tree every time it generated a case, so each case had a number of
<plotfile> elements equal to the number of cases already processed + 1.
This commit fixes the problem by removing all existing <plotfile> elements
before writing the current case's <plotfile> element.

Because FEBio never writes more than one plotfile, it is ok to have only
one <plotfile> element.  The code in this commit will remember the
output variables from the <plotfile> elements it removes and re-add any
unique ones to the new <plotfile> element it adds.