Add Analysis and AnalysisCase classes
The Analysis class stores data defined at the whole-analysis level.  The
AnalysisCase class stores data specific to a particular case.  These
classes were introduced to facilitate use of model generation functions,
which take an AnalysisCase instance as input.

This commit also simplifies the task of using the analysis name for
automatic file naming, as the analysis object is now passed around
ubiquitously.  Before, each function had to figure out the analysis name
by finding the analysis XML file and reading it, which wasn't always
possible.

The Analysis and AnalysisCase objects are meant as input arguments.
They are not modified as the simulations are run and results are
tabulated and plotted.  The tabulation and plotting functions still just
look for files at standard locations in the analysis directory.
Therefore the AnalysisCase objects aren't used after case generation.