Fix handling of FEBio errors
Previously, on an error the captured FEBio stdout stream was written to
the log file.  This is good for file read errors, which would otherwise
result in no log file update.  But it is bad for convergence failures,
for which FEBio's own log file output provides more information.
Therefore, only write stdout to the log file if there is a file read
error.