Avoid confusing mypy
Mypy thinks `ivars_table` is a `defaultdict[str, List[Any]]` even after
it becomes a `DataFrame`.  The easiest workaround is to rename the
DataFrame so each variable has a stable type.

I could instead run mypy with `--allow-redefinition`, but I don't see a
good reason to require special config when it is so easy to avoid type
changes.