Add line plot guide to tsvars heat map figure
Add a line plot to the sensitivity vector heat map to show the time
series variable output for the nominal case (or the median case, if no
nominal case is provided).  The line plot serves as a sort of legend to
show what part of the simulation each step corresponds to.