Use √(unsigned Pearson correlation distance)
If the normal Pearson correlation distance is used, sensitivity vectors
that are opposite in sign but otherwise equal (inverse) will have
distance = 1.  As decreasing one parameter would then have the same
effect as increasing the other, their distance should be 0.

Therefore we use instead a modified distance metric = 1 - | ρ |, where ρ
is the Pearson correlation coefficient.

We apply a square root transform this distance metric to make the
distances distribute more evenly over the available range.  Another way
to look at it is this makes the distance scale like Euclidean distance,
as Euclidean distance = √(2n(1 − ρ)) where n = # data points.  This is
just to make the plots easier to read.