Don't always swallow case generation error
With case generation is run as a batch, usually with do_parallel, the
return code of gen_case is checked.  Thus it makes sense to convert an
FEBioError to return code == 1.  But in interactive use we want to
propagate any exception raised by gen_case, so we should make that the
default .