Store an analysis in a directory named for it
Prior to this commit, if the `directory` argument was provided to
`Analysis.__init__` the analysis name would not be used anywhere, and
the analysis files would simply be written to `directory`.  This was
undesirable.  Either the caller would have to repeat the analysis name
in both `name` and `directory`, switch between providing `name` or
`directory` depending on whether the parent directory should be the
current working directory, or always use `directory` and never `name`.

Now, a folder named `name` is always created to store the analysis
files.  Having the folder named after the analysis makes it clear where
the files came from.  The `directory` argument was renamed `parentdir`
and, as the name implies, spamneggs' behavior was changed so that it
uses `parentdir` / `name` as the analysis directory.