Use variable names directly
Use variable names specified in Analysis object rather than reading a
case's record from disk.  If there are no cases with successful
runs (non-successful runs are filtered out), reading the first record
will of course fail and raise an error.  It's also better to rely on the
analysis definition as the main source of information.