Handle σ_var = 0 in param ∝ var correlations
When computing the correlation coefficient between a parameter and a
variable, the variable may have near-zero variance.  Such is the case if
the variable is totally insensitive to the parameters adjusted in the
sensitivity analysis.  Since the variable's variance is in the
denominator of the Pearson correlation coefficient ρ, this may cause a
result of ρ = NaN.  But in this case the numerator should also be ~ 0,
so we are safe to treat the result as ρ = 0.