Send a filtered cases table to tsvars tabulation
It's not a good idea for tabulate_analysis_tsvars to read the cases
table on its own; it doesn't know whether some cases (with unsuccessful
runs or other issues) should be filtered or not.  Instead, make it take
a table of cases so that the caller can do this filtering ahead of time.