Fix bug in which nan correlations were dropped
NaN correlations were being removed from the parameter–output
correlation vectors.  When points were removed from the start or end of
the time series, the effect was that the heatmaps were the wrong size
and were being stretched to cover the time point axis, such that the
data and the axis no longer lined up.