Fix local cost function plot styling
Previously, the axis offsets (scientific notation scale) were in the
wrong places.  The line plot on the left side of the heatmap had its
x-axis scale on the bottom.  The line plot on the bottom of the heatmap
had its y-axis scale in the usual place on the top left of the axes, but
this collided with the heatmap.

This commit moves all the scales into the axis labels.  It also adds
more accurate axis labels.