Switch to unsigned cosine distances
Standardizing the correlation vectors before calculating the distance
measure means that a parameter with a large mean ρ with tiny
fluctuations will be treated as having ~ zero distance with respect to a
parameter with large fluctuations, as long as they are in phase.  But
these parameters are different; one has a consistent effect over the
whole time series, one affects specific parts of the time series.
Hence, cosine distance, which is a non-standardized version of Pearson
correlation distance.

Also, don't transform the distance by the square root; it makes similar
correlation vectors look too different.