Use hash of parameter vector not base64 encoding
Using the base64-encoding of the parameter vector as the key leads to
long paths that may contain special characters.  Linux handles this
gracefully, but Windows can't deal with it.  And even on Linux, a
forward slash in the bit-packed key produces a subdirectory, which is
kind of weird.

The original reason for using the base64 encoding was to go straight
from the parameter vector to the key.  But using a hash map should work
fine.

FNV-1a was selected because it seems to have a fairly even distribution
over numbers and is simple to implement.  Python used it for a while in
hash(), and only switched because it is vulnerable to "hash
flooding" (deliberately picking values to produce collisions).
EvaluationDB is not meant for use in adversarial environments though.
https://softwareengineering.stackexchange.com/a/145633

I did not use Python's built-in hash() because its key–value
relationship is deliberately randomized each time Python starts, and is
not guaranteed to be consistent across Python versions.